% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_data.R
\docType{data}
\name{data_intersect}
\alias{data_intersect}
\title{Example Data for Intersecting Survival Curves}
\format{
A data frame with 200 rows and 3 variables:
\describe{
  \item{time}{Survival time}
  \item{status}{Survival status (0 = censored, 1 = event)}
  \item{group}{Group variable (Female, Male)}
}
}
\usage{
data_intersect
}
\description{
A dataset containing survival times, status, and groups "Female" and "Male" for survival curves that intersect.
}
\keyword{datasets}
