% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv_curve_intersect.R
\name{surv_curve_intersect}
\alias{surv_curve_intersect}
\title{Check if Survival Curves Intersect}
\usage{
surv_curve_intersect(ggsurvplot_object)
}
\arguments{
\item{ggsurvplot_object}{A ggsurvplot object created by the `survminer` package.}
}
\value{
A logical value indicating if the survival curves intersect (TRUE), do not intersect (FALSE), or if there was an error (NA).
}
\description{
This function checks if two survival curves intersect in a ggsurvplot object.
}
